worker/requirements.txt
alembic==1.18.4
celery[redis]==5.6.3
psycopg2-binary==2.9.11
python-dotenv==1.2.2
redis==6.4.0
sqlalchemy==2.0.48
